Update tests for entities processing
