+++ collect_app/src/main/java/org/odk/collect/android/formmanagement/download/FormDownloadUseCases.kt
@@ -1,12 +1,10 @@
 package org.odk.collect.android.formmanagement.download
 
-import org.javarosa.core.model.instance.CsvExternalInstance
-import org.javarosa.core.model.instance.ExternalDataInstance
 import org.odk.collect.android.formmanagement.ServerFormDetails
 import org.odk.collect.android.utilities.FileUtils
 import org.odk.collect.async.OngoingWorkListener
 import org.odk.collect.entities.EntitiesRepository
-import org.odk.collect.entities.Entity
+import org.odk.collect.entities.LocalEntityUseCases
 import org.odk.collect.forms.Form
 import org.odk.collect.forms.FormSource
 import org.odk.collect.forms.FormSourceException
@@ -75,17 +73,7 @@ object FormDownloadUseCases {
 
             val dataset = mediaFile.filename.substringBefore(".csv")
             if (entitiesRepository.getDatasets().contains(dataset)) {
-                val root = CsvExternalInstance.parse(dataset, tempMediaFile.absolutePath)
-                val items = root.getChildrenWithName("item")
-                items.forEach {
-                    val entity = Entity(
-                        "people",
-                        it.getFirstChild("name")!!.value!!.value as String,
-                        it.getFirstChild("label")!!.value!!.value as String,
-                    )
-
-                    entitiesRepository.save(entity)
-                }
+                LocalEntityUseCases.updateLocalEntities(dataset, tempMediaFile, entitiesRepository)
             }
         }
