+++ collect_app/src/test/java/org/odk/collect/android/formmanagement/download/FormDownloadUseCasesTest.kt
@@ -8,6 +8,7 @@ import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
 import org.odk.collect.android.formmanagement.ServerFormDetails
 import org.odk.collect.android.utilities.FileUtils
+import org.odk.collect.entities.InMemEntitiesRepository
 import org.odk.collect.forms.Form
 import org.odk.collect.forms.FormSource
 import org.odk.collect.forms.ManifestFile
@@ -138,6 +139,7 @@ class FormDownloadUseCasesTest {
             formsRepository,
             File(createTempDir(), "temp").absolutePath,
             createTempDir(),
+            InMemEntitiesRepository(),
             mock()
         )
 
@@ -180,6 +182,7 @@ class FormDownloadUseCasesTest {
             formsRepository,
             File(createTempDir(), "temp").absolutePath,
             createTempDir(),
+            InMemEntitiesRepository(),
             mock()
         )
