+++ entities/src/test/java/org/odk/collect/entities/LocalEntityUseCasesTest.kt
@@ -1,0 +1,27 @@
+package org.odk.collect.entities
+
+import org.junit.Assert.fail
+import org.junit.Test
+
+class LocalEntityUseCasesTest {
+
+    @Test
+    fun `updateLocalEntities overrides local version if the list version is newer`() {
+        fail()
+    }
+
+    @Test
+    fun `updateLocalEntities does not override local version if the list version is older`() {
+        fail()
+    }
+
+    @Test
+    fun `updateLocalEntities does not override local version if the list version is the same`() {
+        fail()
+    }
+
+    @Test
+    fun `updateLocalEntities adds properties not in local version from older list version`() {
+        fail()
+    }
+}
